Move the top score code into a store.
